分布式锁双重防死锁，定时删除关闭订单功能

+++ src/main/Java/com/mmall/service/impl/UserServiceImpl.java
@@ -199,7 +199,7 @@ public class UserServiceImpl implements IUserService {
     //backend
 
     public ServerResponse checkAdminRole(User user){
-        if (user != null && user.getRole().intValue() == Const.Role.ROLE_ADMIN){
+        if (user != null && user.getRole() == Const.Role.ROLE_ADMIN){
             return ServerResponse.createBySuccess();
         }
         return ServerResponse.createByErrorByMessage("非管理员权限");
